Moved client-related elements from BSTBaseListView to BSTClientInterface.
This is a refactor that better encapsulates client interactions in the BSTClientInterface class.  Before, it only defined basic cookie operations.  Now, it houses the template and javascript files, defines the pagination operations and context variable names having to do with the supported BootStrap Table features such as search, filter, and sort, and does the cookie retrieval from the client, regardless of the defined columns.  It also now handles the URL parameters.  The other derived classes are now entirely focussed on column initialization and querying.

Details:

- All but the column objects context variables are handled in get_context.
- Page title (based on the selected model) initializations are now handled here.
- Cookie resets based on client user selections are now handled.
- cookie_warnings was replaced by warnings, previously merged in a derived class.
- Cookie values are set in instance variables in the constructor, for use by derived classes.
- Default model ordering is now handled.
- Pagination class attributes are now here as well.

Files checked in: DataRepo/tests/views/models/bst/test_client_interface.py DataRepo/views/models/bst/base.py DataRepo/views/models/bst/client_interface.py
